pulled automated_target_selection.py and similarity_metrics.py from automated_registration_target_selection branch
